ha-cli: split monolithic ha.ts into types/flags/transport/format + per-group command modules
