+++ bin/ha-cli/eslint.config.mjs
@@ -2,8 +2,9 @@ import globals from 'globals';
 import tseslint from 'typescript-eslint';
 
 export default tseslint.config(
+  { ignores: ['node_modules/**'] },
   {
-    files: ['ha.ts'],
+    files: ['**/*.ts'],
     languageOptions: {
       parserOptions: {
         projectService: true,
